refactor entities, repository methods and openweathermap plugin

+++ WeatherParser.Presentation.Entities/WeatherPresentation.cs
@@ -10,7 +10,7 @@ namespace WeatherParser.Presentation.Entities
         public List<double> Temperature { get; set; }
         public List<int> Pressure { get; set; }
         public List<int> Humidity { get; set; }
-        public List<int> WindSpeed { get; set; }
+        public List<double> WindSpeed { get; set; }
         public List<string> WindDirection { get; set; }
 
     }
