+++ WeatherParser.Repository/Entities/WeatherRepository.cs
@@ -14,7 +14,7 @@ namespace WeatherParser.Repository.Entities
         public List<double> Temperature { get; set; }
         public List<int> Pressure { get; set; }
         public List<int> Humidity { get; set; }
-        public List<int> WindSpeed { get; set; }
+        public List<double> WindSpeed { get; set; }
         public List<string> WindDirection { get; set; }
     }
 }
