+++ WeatherParser.Service.Entities/WeatherService.cs
@@ -7,7 +7,7 @@
         public List<double> Temperature { get; set; }
         public List<int> Pressure { get; set; }
         public List<int> Humidity { get; set; }
-        public List<int> WindSpeed { get; set; }
+        public List<double> WindSpeed { get; set; }
         public List<string> WindDirection { get; set; }
 
     }
