+++ WeatherParser.Service.ForecaService/WeatherDataAngleSharpServiceForeca.cs
@@ -42,7 +42,7 @@ namespace WeatherParser.Service.Plugins.ForecaService
                     Humidity = new List<int>(),
                     Pressure = new List<int>(),
                     WindDirection = new List<string>(),
-                    WindSpeed = new List<int>()
+                    WindSpeed = new List<double>()
                 };
 
                 //TODO PARSE DATE
